// SGUEES-SPA/src/app/pages/Shop/com-proveedor-actu/models/com-proveedor-actu.ts
export interface ComProveedorActu {
	CORR_PROVEEDOR: number;
	CODIGO_PROVEEDOR: string;
	TIPO_PERSONERIA: string;
	NOMBRE_TIPO_PERSONERIA: string;
	NOMBRE_PROVEEDOR: string;
	PRIMER_NOMBRE: string;
	SEGUNDO_NOMBRE: string;
	PRIMER_APELLIDO: string;
	SEGUNDO_APELLIDO: string;
	NOMBRE_COMERCIAL: string;
	CORR_TIPO_DIP: number;
	NOMBRE_TIPO_DIP: string;
	NUMERO_DIP: string;
	NUMERO_NRC: string;
	NUMERO_NIT: string;
	CORR_ACTIVIDAD_ECONOMICA: string;
	NOMBRE_ACTIVIDAD_ECONOMICA: string;
	DIRECCION_PROVEEDOR: string;
	CODIGO_PAIS: string;
	NOMBRE_PAIS: string;
	CODIGO_DEPTO: string;
	NOMBRE_DEPTO: string;
	CODIGO_MUNICIPIO: string;
	NOMBRE_MUNICIPIO: string;
	NOMBRE_CONTACTO: string;
	TELEFONO_FIJO: string;
	TELEFONO_MOVIL: string;
	CORREO_ELECTRONICO_1: string;
	CORREO_ELECTRONICO_2: string;
	CORR_FORMA_PAGO: number;
	NOMBRE_FORMA_PAGO: string;
	CUENTA_BANCARIA: string;
	CORR_BANCO: number;
	NOMBRE_BANCO: string;
	ESTADO_PROVEEDOR: string;
	NOMBRE_ESTADO_PROVEEDOR: string;
	ESTADO_PROVEEDOR_WEB: string;
	NOMBRE_ESTADO_PROVEEDOR_WEB: string;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
}